Generic sweeps for floating gate experiments
This commit contains a couple generic sweep functions that were
used to measure the floating gate structures in the dunker.

'time_sweep' can take a list of 'events' and play them sequentially,
measuring the results over time.

'linear_sweep' is a simple linear sweep of one parameter and a
number of dependents.